close #14

+++ genieclust/c_genie.pxd
@@ -47,7 +47,7 @@ cdef extern from "../src/c_genie.h":
     cdef cppclass CGenie[T]:
         CGenie() except +
         CGenie(T* mst_d, ssize_t* mst_i, ssize_t n, bint noise_leaves) except +
-        void apply_genie(ssize_t n_clusters, double gini_threshold, int* res)
+        void apply_genie(ssize_t n_clusters, double gini_threshold, ssize_t* res)
         void apply_gic(ssize_t n_clusters, ssize_t add_clusters,
             double n_features,
-            double* gini_thresholds, ssize_t n_thresholds, int* res)
+            double* gini_thresholds, ssize_t n_thresholds, ssize_t* res)
